修复migrations

+++ database/migrations/2017_12_15_145350_create_ims_yz_member_address_table.php
@@ -12,21 +12,23 @@ class CreateImsYzMemberAddressTable extends Migration {
 	 */
 	public function up()
 	{
-		Schema::create('yz_member_address', function(Blueprint $table)
-		{
-			$table->increments('id');
-			$table->integer('uniacid')->unsigned()->index('idx_uinacid');
-			$table->integer('uid')->unsigned()->index('idx_uid');
-			$table->string('username', 20);
-			$table->string('mobile', 11);
-			$table->string('zipcode', 6);
-			$table->string('province', 32);
-			$table->string('city', 32);
-			$table->string('district', 32);
-			$table->string('street', 32);
-			$table->string('address', 512);
-			$table->boolean('isdefault');
-		});
+        if (!Schema::hasTable('yz_member_address')) {
+
+            Schema::create('yz_member_address', function (Blueprint $table) {
+                $table->increments('id');
+                $table->integer('uniacid')->unsigned()->index('idx_uinacid');
+                $table->integer('uid')->unsigned()->index('idx_uid');
+                $table->string('username', 20);
+                $table->string('mobile', 11);
+                $table->string('zipcode', 6);
+                $table->string('province', 32);
+                $table->string('city', 32);
+                $table->string('district', 32);
+                $table->string('street', 32);
+                $table->string('address', 512);
+                $table->boolean('isdefault');
+            });
+        }
 	}
 
 
@@ -37,7 +39,10 @@ class CreateImsYzMemberAddressTable extends Migration {
 	 */
 	public function down()
 	{
-		Schema::drop('ims_yz_member_address');
+        if (Schema::hasTable('ims_yz_member_address')) {
+
+            Schema::drop('ims_yz_member_address');
+        }
 	}
 
 }
